dfir_rs::compiled

Module pull

Pull-based operator helpers, i.e. [futures::Stream] helpers.

Structs

CrossSingleton

Stream combinator that crosses each item from item_stream with a singleton value from singleton_stream.

Flatten

Same as Iterator::flatten but as a [Stream].

ForEach

A future which consumes a stream by feeding a sync function Func with all items.

HalfMultisetJoinState

HalfJoinState with multiset semantics.

HalfSetJoinState

HalfJoinState with set semantics.

IntoNext

A future which resolves with the next item in the stream.

LatticeBimorphismStream

Stream combinator for lattice bimorphism operations.

SymmetricHashJoin

Stream combinator for symmetric hash join operations.

ZipLongest

Special stream for the zip_longest operator.

Traits

HalfJoinState

State semantics for each half of a join.

Functions

accumulate_all

Use the accumulator accum to accumulate all entries in the stream st into the hash_map.

symmetric_hash_join_into_stream

Creates a symmetric hash join stream from two input streams and their join states.