dfir_rs

Module scheduled

DFIR’s outer scheduled layer. Deals with inter-subgraph runtime data-passing and scheduling.

The most important item is the Dfir struct. Most of the items in this module are supporting the implementation of that struct and its operation.

Modules

context

Module for the user-facing Context object.

graph

Module for the Dfir struct and helper items.

graph_ext

Helper extensions for Dfir.

handoff

Module for all Handoff-related items.

input

metrics

Runtime metrics for DFIR.

net

This module previously contained networking code.

port

Organizational module for SendCtx/RecvCtx structs and SendPort/RecvPort structs.

reactor

Module for Reactor.

state

Module for StateHandle, part of the “state API”.

ticks

This module contains types to work with ticks.

Enums

HandoffTag

Tag for HandoffId.

LoopTag

Tag for LoopId.

StateTag

Tag for StateId.

SubgraphTag

Tag for SubgraphId.

Type Aliases

HandoffId

A handoff’s ID. Invalid if used in a different graph::Dfir instance than the original that created it.

LoopId

A loop’s ID.

StateId

A staten handle’s ID. Invalid if used in a different graph::Dfir instance than the original that created it.

SubgraphId

A subgraph’s ID. Invalid if used in a different graph::Dfir instance than the original that created it.