dfir_rs::scheduled::context

Struct Context

pub struct Context { /* private fields */ }

The main state and scheduler of the runtime instance. Provided as the context API to each subgraph/operator as it is run.

Each instance stores eactly one Context inline. Before the Context is provided to a running operator, the subgraph_id field must be updated.

Implementations

impl Context

Public APIs.

pub fn current_tick(&self) -> TickInstant

Gets the current tick (local time) count.

pub fn current_tick_start(&self) -> SystemTime

Gets the timestamp of the beginning of the current tick.

pub fn is_first_run_this_tick(&self) -> bool

Gets whether this is the first time this subgraph is being scheduled for this tick

pub fn loop_iter_count(&self) -> usize

Gets the current loop iteration count.

pub fn current_stratum(&self) -> usize

Gets the current stratum nubmer.

pub fn current_subgraph(&self) -> SubgraphId

Gets the ID of the current subgraph.

pub fn schedule_subgraph(&self, sg_id: SubgraphId, is_external: bool)

Schedules a subgraph for the next tick.

If is_external is true, the scheduling will trigger the next tick to begin. If it is false then scheduling will be lazy and the next tick will not begin unless there is other reason to.

pub fn reschedule_loop_block(&self)

Schedules the current loop block to be run again (in this tick).

pub fn allow_another_iteration(&self)

Allow another iteration of the loop, if more data comes.

pub fn waker(&self) -> Waker

Returns a Waker for interacting with async Rust. Waker events are considered to be extenral.

pub unsafe fn state_ref_unchecked<T>(&self, handle: StateHandle<T>) -> &T
where T: Any,

Returns a shared reference to the state.

Safety

StateHandle<T> must be from this instance, created via Self::add_state.

pub fn state_ref<T>(&self, handle: StateHandle<T>) -> &T
where T: Any,

Returns a shared reference to the state.

pub fn state_mut<T>(&mut self, handle: StateHandle<T>) -> &mut T
where T: Any,

Returns an exclusive reference to the state.

pub fn add_state<T>(&mut self, state: T) -> StateHandle<T>
where T: Any,

Adds state to the context and returns the handle.

pub fn set_state_lifespan_hook<T>( &mut self, handle: StateHandle<T>, lifespan: StateLifespan, hook_fn: impl 'static + FnMut(&mut T), )
where T: Any,

Sets a hook to modify the state at the end of each tick, using the supplied closure.

pub fn request_task<Fut>(&mut self, future: Fut)
where Fut: Future<Output = ()> + 'static,

Prepares an async task to be launched by Self::spawn_tasks.

pub fn spawn_tasks(&mut self)

Launches all tasks requested with Self::request_task on the internal Tokio executor.

pub fn abort_tasks(&mut self)

Aborts all tasks spawned with Self::spawn_tasks.

pub async fn join_tasks(&mut self)

Waits for all tasks spawned with Self::spawn_tasks to complete.

Will probably just hang.