dfir_rs::scheduled::graph

Struct Dfir

pub struct Dfir<'a> { /* private fields */ }

A DFIR graph. Owns, schedules, and runs the compiled subgraphs.

Implementations

impl Dfir<'_>

Methods for TeeingHandoff teeing and dropping.

pub fn teeing_handoff_tee<T>( &mut self, tee_parent_port: &RecvPort<TeeingHandoff<T>>, ) -> RecvPort<TeeingHandoff<T>>
where T: Clone,

Tees a TeeingHandoff.

pub fn teeing_handoff_drop<T>(&mut self, tee_port: RecvPort<TeeingHandoff<T>>)
where T: Clone,

Marks an output of a TeeingHandoff as dropped so that no more data will be sent to it.

It is recommended to not not use this method and instead simply avoid teeing a TeeingHandoff when it is not needed.

impl<'a> Dfir<'a>

pub fn new() -> Self

Create a new empty graph.

pub fn meta_graph(&self) -> Option<&DfirGraph>

Available on crate feature meta only.

Return a handle to the meta graph, if set. The meta graph is a representation of all the operators, subgraphs, and handoffs in this instance. Will only be set if this graph was constructed using a surface syntax macro.

pub fn diagnostics(&self) -> Option<&[Diagnostic<SerdeSpan>]>

Available on crate feature meta only.

Returns any diagnostics generated by the surface syntax macro. Each diagnostic is a pair of (1) a Diagnostic with span info reset and (2) the ToString version of the diagnostic with original span info. Will only be set if this graph was constructed using a surface syntax macro.

pub fn reactor(&self) -> Reactor

Returns a reactor for externally scheduling subgraphs, possibly from another thread. Reactor events are considered to be external events.

pub fn current_tick(&self) -> TickInstant

Gets the current tick (local time) count.

pub fn current_stratum(&self) -> usize

Gets the current stratum nubmer.

pub async fn run_tick(&mut self) -> bool

Runs the dataflow until the next tick begins.

Returns true if any work was done.

Will receive events if it is the start of a tick when called.

pub fn run_tick_sync(&mut self) -> bool

Self::run_tick but panics if a subgraph yields asynchronously.

pub async fn run_available(&mut self) -> bool

Runs the dataflow until no more (externally-triggered) work is immediately available. Runs at least one tick of dataflow, even if no external events have been received. If the dataflow contains loops this method may run forever.

Returns true if any work was done.

Yields repeatedly to allow external events to happen.

pub fn run_available_sync(&mut self) -> bool

Self::run_available but panics if a subgraph yields asynchronously.

pub async fn run_stratum(&mut self) -> bool

Runs the current stratum of the dataflow until no more local work is available (does not receive events).

Returns true if any work was done.

pub fn next_stratum(&mut self, current_tick_only: bool) -> bool

Go to the next stratum which has work available, possibly the current stratum. Return true if more work is available, otherwise false if no work is immediately available on any strata.

This will receive external events when at the start of a tick.

If current_tick_only is set to true, will only return true if work is immediately available on the current tick.

If this returns false then the graph will be at the start of a tick (at stratum 0, can receive more external events).

pub async fn run(&mut self) -> Option<Never>

Runs the dataflow graph forever.

TODO(mingwei): Currently blocks forever, no notion of “completion.”

pub fn run_sync(&mut self) -> Option<Never>

Self::run but panics if a subgraph yields asynchronously.

pub fn try_recv_events(&mut self) -> usize

Enqueues subgraphs triggered by events without blocking.

Returns the number of subgraphs enqueued, and if any were external.

pub fn recv_events(&mut self) -> Option<usize>

Enqueues subgraphs triggered by external events, blocking until at least one subgraph is scheduled from an external event.

pub async fn recv_events_async(&mut self) -> Option<usize>

Enqueues subgraphs triggered by external events asynchronously, waiting until at least one subgraph is scheduled from an external event. Returns the number of subgraphs enqueued, which may be zero if an external event scheduled an already-scheduled subgraph.

Returns None if the event queue is closed, but that should not happen normally.

pub fn schedule_subgraph(&mut self, sg_id: SubgraphId) -> bool

Schedules a subgraph to be run. See also: Context::schedule_subgraph.

pub fn add_subgraph<Name, R, W, Func>( &mut self, name: Name, recv_ports: R, send_ports: W, subgraph: Func, ) -> SubgraphId
where Name: Into<Cow<'static, str>>, R: 'static + PortList<RECV>, W: 'static + PortList<SEND>, Func: 'a + for<'ctx> AsyncFnMut(&'ctx mut Context, R::Ctx<'ctx>, W::Ctx<'ctx>),

Adds a new compiled subgraph with the specified inputs and outputs in stratum 0.

pub fn add_subgraph_stratified<Name, R, W, Func>( &mut self, name: Name, stratum: usize, recv_ports: R, send_ports: W, laziness: bool, subgraph: Func, ) -> SubgraphId
where Name: Into<Cow<'static, str>>, R: 'static + PortList<RECV>, W: 'static + PortList<SEND>, Func: 'a + for<'ctx> AsyncFnMut(&'ctx mut Context, R::Ctx<'ctx>, W::Ctx<'ctx>),

Adds a new compiled subgraph with the specified inputs, outputs, and stratum number.

TODO(mingwei): add example in doc.

pub fn add_subgraph_full<Name, R, W, Func>( &mut self, name: Name, stratum: usize, recv_ports: R, send_ports: W, laziness: bool, loop_id: Option<LoopId>, subgraph: Func, ) -> SubgraphId
where Name: Into<Cow<'static, str>>, R: 'static + PortList<RECV>, W: 'static + PortList<SEND>, Func: 'a + for<'ctx> AsyncFnMut(&'ctx mut Context, R::Ctx<'ctx>, W::Ctx<'ctx>),

Adds a new compiled subgraph with all options.

pub fn add_subgraph_n_m<Name, R, W, Func>( &mut self, name: Name, recv_ports: Vec<RecvPort<R>>, send_ports: Vec<SendPort<W>>, subgraph: Func, ) -> SubgraphId
where Name: Into<Cow<'static, str>>, R: 'static + Handoff, W: 'static + Handoff, Func: 'a + for<'ctx> AsyncFnMut(&'ctx mut Context, &'ctx [&'ctx RecvCtx<R>], &'ctx [&'ctx SendCtx<W>]),

Adds a new compiled subgraph with a variable number of inputs and outputs of the same respective handoff types.

pub fn add_subgraph_stratified_n_m<Name, R, W, Func>( &mut self, name: Name, stratum: usize, recv_ports: Vec<RecvPort<R>>, send_ports: Vec<SendPort<W>>, subgraph: Func, ) -> SubgraphId
where Name: Into<Cow<'static, str>>, R: 'static + Handoff, W: 'static + Handoff, Func: 'a + for<'ctx> AsyncFnMut(&'ctx mut Context, &'ctx [&'ctx RecvCtx<R>], &'ctx [&'ctx SendCtx<W>]),

Adds a new compiled subgraph with a variable number of inputs and outputs of the same respective handoff types.

pub fn make_edge<Name, H>(&mut self, name: Name) -> (SendPort<H>, RecvPort<H>)
where Name: Into<Cow<'static, str>>, H: 'static + Handoff,

Creates a handoff edge and returns the corresponding send and receive ports.

pub fn add_state<T>(&mut self, state: T) -> StateHandle<T>
where T: Any,

Adds referenceable state into this instance. Returns a state handle which can be used externally or by operators to access the state.

This is part of the “state API”.

pub fn set_state_lifespan_hook<T>( &mut self, handle: StateHandle<T>, lifespan: StateLifespan, hook_fn: impl 'static + FnMut(&mut T), )
where T: Any,

Sets a hook to modify the state at the end of each tick, using the supplied closure.

This is part of the “state API”.

pub fn context_mut(&mut self, sg_id: SubgraphId) -> &mut Context

Gets a exclusive (mut) ref to the internal context, setting the subgraph ID.

pub fn add_loop(&mut self, parent: Option<LoopId>) -> LoopId

Adds a new loop with the given parent (or None for top-level). Returns a loop ID which is used in Self::add_subgraph_stratified or for nested loops.

TODO(mingwei): add loop names to ensure traceability while debugging?

pub fn metrics(&self) -> Rc<DfirMetrics>

Returns a reference-counted handle to the continually-updated runtime metrics for this DFIR instance.

pub fn metrics_intervals(&self) -> DfirMetricsIntervals

Returns a DfirMetricsIntervals handle where each call to DfirMetricsIntervals::take_interval ends the current interval and returns its metrics.

The first call to take_interval returns metrics since this DFIR instance was created. Each subsequent call to take_interval returns metrics since the previous call.

Cloning the handle “forks” it from the original, as afterwards each interval may return different metrics depending on when exactly take_interval is called.

impl Dfir<'_>

pub fn request_task<Fut>(&mut self, future: Fut)
where Fut: Future<Output = ()> + 'static,

Alias for Context::request_task.

pub fn abort_tasks(&mut self)

Alias for Context::abort_tasks.

pub fn join_tasks(&mut self) -> impl use<'_> + Future

Alias for Context::join_tasks.