dfir_rs::scheduled::handoff

Module handoff_list

Module for variadic handoff port lists, PortList.

Traits

HandoffList

A variadic list of Handoff types, represented using a lisp-style tuple structure.

PortList

Sealed trait for variadic lists of ports.

PortListSplit

Trait for splitting a list of ports into two.