dfir_rs::scheduled::port

Type Alias RecvPort

pub type RecvPort<H> = Port<RECV, H>;

Recv-specific variant of Port. An input port.

Aliased Type

pub struct RecvPort<H> { /* private fields */ }

Implementations

impl<T: Clone> RecvPort<TeeingHandoff<T>>

Methods for TeeingHandoff teeing and dropping.

pub fn tee(&self, hf: &mut Dfir<'_>) -> RecvPort<TeeingHandoff<T>>

Tees this TeeingHandoff, given the Dfir instance it belongs to.

pub fn drop(self, hf: &mut Dfir<'_>)

Marks this output of a TeeingHandoff as dropped so that no more data will be sent to it, given the Dfir instance it belongs to.

It is recommended to not not use this method and instead simply avoid teeing a TeeingHandoff when it is not needed.