dfir_rs::util::slot_vec

Struct SlotVec

pub struct SlotVec<Tag: ?Sized, Val> { /* private fields */ }

A Vec-based SlotMap-esque datastructure without removes.

Analogous to slotmap::SlotMap, but avoids the overhead of tracking removed keys.

Implementations

impl<Tag: ?Sized, Val> SlotVec<Tag, Val>

pub const fn new() -> Self

Creates a new SlotVec.

pub fn len(&self) -> usize

Returns the number of elements in the SlotVec.

pub fn is_empty(&self) -> bool

Returns true if the SlotVec is empty.

pub fn insert(&mut self, value: Val) -> Key<Tag>

Inserts a value into the SlotVec and returns the key.

pub fn insert_with_key<F>(&mut self, func: F) -> Key<Tag>
where F: FnOnce(Key<Tag>) -> Val,

Use the provided function to generate a value given the key and insert it into the SlotVec.

pub fn get(&self, key: Key<Tag>) -> Option<&Val>

Returns a reference to the value associated with the key.

pub fn get_mut(&mut self, key: Key<Tag>) -> Option<&mut Val>

Returns a mutable reference to the value associated with the key.

pub fn iter( &self, ) -> impl DoubleEndedIterator<Item = (Key<Tag>, &Val)> + ExactSizeIterator + FusedIterator + Clone

Iterate the key-value pairs, where the value is a shared reference.

pub fn iter_mut( &mut self, ) -> impl DoubleEndedIterator<Item = (Key<Tag>, &mut Val)> + ExactSizeIterator + FusedIterator

Iterate the key-value pairs, where the value is a exclusive reference.

pub fn values(&self) -> Iter<'_, Val>

Iterate over the values by shared reference.

pub fn values_mut(&mut self) -> IterMut<'_, Val>

Iterate over the values by exclusive reference.

pub fn keys( &self, ) -> impl '_ + DoubleEndedIterator<Item = Key<Tag>> + ExactSizeIterator + FusedIterator + Clone

Iterate over the keys.